order-service/pyproject.toml
[project]
name = "order-service"
version = "1.0.0"
description = ""
authors = [
    {name = "Caio POrcel",email = "caio.porcel@cs.up.edu.br"}
]
requires-python = ">=3.12"
dependencies = [
    "flask (>=3.1.1,<4.0.0)",
    "redis (>=6.2.0,<7.0.0)",
    "mysql-connector-python (>=9.3.0,<10.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]

[tool.poetry]
name = "order-service"
version = "1.0.0"
packages = [{ include = "src" }]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "src:main"